# mcp_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Notion MCP Governance-First Server
  version: 1.4.0
  description: |
    MCP server for Notion with governance, workflow, audit, and approval tools.
    Mirrors official Notion MCP tool schema and supports the new data-source API.
servers:
  - url: http://localhost:8080
paths:
...
  /call_tool/policy.check:
    post:
      summary: Check governance policies for an action
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCheckInput'
      responses:
        '200':
          description: Policy check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyCheckResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/MissingField'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /rag_tool/rag.retrieve:
    post:
      summary: Governance-first vector retrieval
      security:
        - apiKeyAuth: []
      description: |
        Filters at source level (workspace_id resolved from API key, status='published',
        effective window, legal_hold=FALSE) before ranking chunks by cosine similarity.
        Optionally scopes to a specific child via child_id (joins rag_source_access).
        workspace_id is NOT accepted in the request body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagRetrieveInput'
      responses:
        '200':
          description: Ranked chunks with similarity scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RagRetrieveResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/MissingField'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /rag_tool/resource.list:
    post:
      summary: List authorized resources via secure SQL view
      security:
        - apiKeyAuth: []
      description: |
        Standalone, stateless authorized resource listing.
        The caller supplies actor identity per request, the server resolves workspace_id
        from X-API-Key, and SQL authorization is enforced through secure views.
        Only curated resource fields are returned; no raw payload persistence is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceListInput'
      responses:
        '200':
          description: Authorized resources visible to the supplied actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResult'
        '401':
          description: Missing or invalid X-API-Key
        '422':
          $ref: '#/components/responses/MissingField'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/ServerError'
  /rag_tool/resource.get:
    post:
      summary: Get one authorized resource via secure SQL view
      security:
        - apiKeyAuth: []
      description: |
        Standalone, stateless authorized resource lookup.
        The caller supplies actor identity per request, the server resolves workspace_id
        from X-API-Key, and SQL authorization is enforced through secure views.
        Absent or unauthorized resources return null to avoid existence disclosure.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceGetInput'
      responses:
        '200':
          description: Authorized resource if visible to the supplied actor; null otherwise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGetResult'
        '401':
          description: Missing or invalid X-API-Key
        '422':
          $ref: '#/components/responses/MissingField'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Required for all tool endpoints (both /call_tool/* and /rag_tool/*).
        Value is hashed with HMAC-SHA256 (using RAG_SERVER_SECRET) and looked up in RAG_API_KEYS.
        The matching workspace_id is injected server-side via UUIDv5 derivation; never pass it in the body.
        Rate limited to 60 requests/minute per API key.
  responses:
    MissingField:
      description: Missing required field in request body
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    RateLimitExceeded:
      description: Rate limit exceeded (60 requests/minute per API key)
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Rate limit exceeded. Maximum 60 requests per minute per API key."
      headers:
        Retry-After:
          schema:
            type: integer
            example: 60
          description: Number of seconds to wait before retrying
    Unauthorized:
      description: Missing or invalid X-API-Key header
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Missing or invalid API key"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
  schemas:
    PolicyCheckInput:
      type: object
      required:
        - actor
        - action
      properties:
        actor:
          type: string
        action:
          type: string
        context:
          type: object
    PolicyCheckResult:
      type: object
      properties:
        matchedPolicies:
          type: integer
...
    RagRetrieveInput:
      type: object
      required:
        - query_embedding
      properties:
        query_embedding:
          type: array
          items:
            type: number
          description: Embedding vector of the search query.
        top_k:
          type: integer
          default: 10
        child_id:
          type: string
          description: Optional. When set, adds a join on rag_source_access.
    RagRetrieveResult:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:              { type: string }
              source_id:       { type: string }
              content:         { type: string }
              position:        { type: integer }
              token_count:     { type: integer }
              source_name:     { type: string }
              visibility:      { type: string }
              retention_class: { type: string }
              similarity:      { type: number }
...
    ResourceListInput:
      type: object
      required:
        - actorId
      properties:
        actorId:
          type: string
        actorType:
          type: string
          default: user
        memberId:
          type: string
        resourceType:
          type: string
        limit:
          type: integer
          default: 50
        requestId:
          type: string
    ResourceGetInput:
      type: object
      required:
        - actorId
        - resourceId
      properties:
        actorId:
          type: string
        actorType:
          type: string
          default: user
        resourceId:
          type: string
        requestId:
          type: string
    ResourceRecord:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        member_id:
          type: string
          nullable: true
        source_id:
          type: string
          nullable: true
        resource_type:
          type: string
        name:
          type: string
        status:
          type: string
        visibility:
          type: string
        classification:
          type: string
          nullable: true
        effective_from:
          type: string
          nullable: true
        effective_to:
          type: string
          nullable: true
        retention_class:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
    ResourceListResult:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceRecord'
    ResourceGetResult:
      type: object
      properties:
        resource:
          allOf:
            - $ref: '#/components/schemas/ResourceRecord'
          nullable: true